+++ src/webapp/vue.config.js
@@ -40,7 +40,8 @@ module.exports = {
       new webpack.EnvironmentPlugin({
         MELTANO_APP_URL: 'http://localhost:5000',
         MELTANO_OAUTH_SERVICE_URL: 'http://localhost:5000/-/oauth',
-        MELTANO_OAUTH_SERVICE_PROVIDERS: 'all'
+        MELTANO_OAUTH_SERVICE_PROVIDERS: 'all',
+        MELTANO_NOTIFICATION: 'False'
       })
     ]
   },