+++ src/webapp/src/api/design.js
@@ -18,6 +18,10 @@ export default {
     return axios.get(utils.apiUrl('repos/tables', `${table}`))
   },
 
+  getTopic(namespace, model) {
+    return axios.get(utils.apiUrl('repos', `${namespace}/${model}`))
+  },
+
   index(namespace, model, design) {
     return axios.get(
       utils.apiUrl('repos/designs', `${namespace}/${model}/${design}`)