+++ CHANGELOG.md
@@ -9,7 +9,11 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 
+- [#1601](https://gitlab.com/meltano/meltano/issues/1601) Add Explore landing pages per data source to act as an aggregate jump-off point to related dashboards, reports, report templates, and more
+
 ### Changes
+
+- [#1601](https://gitlab.com/meltano/meltano/issues/1601) Change "Reports" CTA in each Pipeline and the JobLog modal to link to its corresponding and newly added Explore page
 
 ### Fixes